requirements.txt
Pillow
opencv-python
tkinter
Flask
go-mtpfs
PySimpleGUI